VGG16 FCN 32s - resul, loss: 0.28
Input image: 224 x 224
Final log output:

Global epoch: 60/100
Epoch 1/5
40/40 [==============================] - 26s - loss: 0.2648 - acc: 0.9089
Epoch 2/5
40/40 [==============================] - 26s - loss: 0.2578 - acc: 0.9114
Epoch 3/5
40/40 [==============================] - 26s - loss: 0.2549 - acc: 0.9125
Epoch 4/5
40/40 [==============================] - 26s - loss: 0.2570 - acc: 0.9109
Epoch 5/5
40/40 [==============================] - 26s - loss: 0.2553 - acc: 0.9124
Validation data mean loss:
Mean loss: 0.282585293055 - mean acc: 0.901066675782

+++ data_handling.py
@@ -244,8 +244,8 @@ def load_data_from_npz(file_name_list):
         sys.exit("ERROR: File name list empty.")
 
     loaded_data = np.load(file_name_list[0])
-    image = loaded_data["image"]
-    label = loaded_data["label"]
+    image = loaded_data["image"].astype(np.float32)
+    label = loaded_data["label"].astype(np.float32)
 
     iw, ih, ic = image.shape
     mw, mh, mc = label.shape
@@ -258,8 +258,8 @@ def load_data_from_npz(file_name_list):
 
     for i in range(1, n_files):
         loaded_data = np.load(file_name_list[i])
-        image = loaded_data["image"]
-        label = loaded_data["label"]
+        image = loaded_data["image"].astype(np.float32)
+        label = loaded_data["label"].astype(np.float32)
 
         x_data[i, :, :, :] = image
         y_data[i, :, :, :] = label
